--- a/senders/formatter.go
+++ b/senders/formatter.go
@@ -13,7 +13,6 @@ import (
 // <metricName> <metricValue> [<timestamp>] source=<source> [pointTags]
 // Example: "new-york.power.usage 42422.0 1533531013 source=localhost datacenter=dc1"
 func MetricLine(name string, value float64, ts int64, source string, tags map[string]string, defaultSource string) (string, error) {
-
 	if name == "" {
 		return "", errors.New("empty metric name")
 	}
@@ -22,7 +21,9 @@ func MetricLine(name string, value float64, ts int64, source string, tags map[st
 		source = defaultSource
 	}
 
-	sb := bytes.Buffer{}
+	sb := GetBuffer()
+	defer PutBuffer(sb)
+
 	sb.WriteString(strconv.Quote(name))
 	sb.WriteString(" ")
 	sb.WriteString(strconv.FormatFloat(value, 'f', -1, 64))
@@ -52,7 +53,6 @@ func MetricLine(name string, value float64, ts int64, source string, tags map[st
 // {!M | !H | !D} [<timestamp>] #<count> <mean> [centroids] <histogramName> source=<source> [pointTags]
 // Example: "!M 1533531013 #20 30.0 #10 5.1 request.latency source=appServer1 region=us-west"
 func HistoLine(name string, centroids []histogram.Centroid, hgs map[histogram.Granularity]bool, ts int64, source string, tags map[string]string, defaultSource string) (string, error) {
-
 	if name == "" {
 		return "", errors.New("empty distribution name")
 	}
@@ -69,12 +69,14 @@ func HistoLine(name string, centroids []histogram.Centroid, hgs map[histogram.Gr
 		source = defaultSource
 	}
 
-	// Preprocess line. We know len(hgs) > 0 here.
-	sb := bytes.Buffer{}
+	sb := GetBuffer()
+	defer PutBuffer(sb)
+
 	if ts != 0 {
 		sb.WriteString(" ")
 		sb.WriteString(strconv.FormatInt(ts, 10))
 	}
+	// Preprocess line. We know len(hgs) > 0 here.
 	for _, centroid := range centroids {
 		sb.WriteString(" #")
 		sb.WriteString(strconv.Itoa(centroid.Count))
@@ -112,7 +114,6 @@ func HistoLine(name string, centroids []histogram.Centroid, hgs map[histogram.Gr
 // "getAllUsers source=localhost traceId=7b3bf470-9456-11e8-9eb6-529269fb1459 spanId=0313bafe-9457-11e8-9eb6-529269fb1459
 //    parent=2f64e538-9457-11e8-9eb6-529269fb1459 application=Wavefront http.method=GET 1533531013 343500"
 func SpanLine(name string, startMillis, durationMillis int64, source, traceId, spanId string, parents, followsFrom []string, tags []SpanTag, spanLogs []SpanLog, defaultSource string) (string, error) {
-
 	if name == "" {
 		return "", errors.New("empty span name")
 	}
@@ -128,7 +129,9 @@ func SpanLine(name string, startMillis, durationMillis int64, source, traceId, s
 		return "", errors.New("spanId is not in UUID format")
 	}
 
-	sb := bytes.Buffer{}
+	sb := GetBuffer()
+	defer PutBuffer(sb)
+
 	sb.WriteString(strconv.Quote(name))
 	sb.WriteString(" source=")
 	sb.WriteString(strconv.Quote(source))