Improves memory use/allocations when building point lines.

Metrics:
```
benchmark                 old ns/op     new ns/op     delta
BenchmarkMetricLine-8     887           925           +4.28%

benchmark                 old allocs     new allocs     delta
BenchmarkMetricLine-8     10             9              -10.00%

benchmark                 old bytes     new bytes     delta
BenchmarkMetricLine-8     232           168           -27.59%
```

Histograms:
```
benchmark                old ns/op     new ns/op     delta
BenchmarkHistoLine-8     1499          1403          -6.40%

benchmark                old allocs     new allocs     delta
BenchmarkHistoLine-8     15             13             -13.33%

benchmark                old bytes     new bytes     delta
BenchmarkHistoLine-8     728           520           -28.57%
```

Spans:
```
benchmark               old ns/op     new ns/op     delta
BenchmarkSpanLine-8     955           768           -19.58%

benchmark               old allocs     new allocs     delta
BenchmarkSpanLine-8     8              5              -37.50%

benchmark               old bytes     new bytes     delta
BenchmarkSpanLine-8     880           256           -70.91%
```